+++ src/cells/LongNumberCell.cpp
@@ -265,15 +265,6 @@ bool LongNumberCell::BreakUp() const {
       }
     }
   }
-  m_innerCell->last()->SetNextToDraw(m_nextToDraw);
-  m_nextToDraw = m_innerCell;
   Cell::BreakUpAndMark();
   return true;
-}
-
-void LongNumberCell::SetNextToDraw(Cell *next) const {
-  if (IsBrokenIntoLines())
-    m_innerCell->last()->SetNextToDraw(next);
-  else
-    m_nextToDraw = next;
 }