+++ src/cells/ParenCell.cpp
@@ -347,21 +347,5 @@ bool ParenCell::BreakUp() const {
     return false;
 
   Cell::BreakUpAndMark();
-  if(m_innerCell)
-    {
-      m_open->SetNextToDraw(m_innerCell);
-      m_innerCell->last()->SetNextToDraw(m_close);
-    }
-  else
-    m_open->SetNextToDraw(m_close);
-  m_close->SetNextToDraw(m_nextToDraw);
-  m_nextToDraw = m_open;
   return true;
-}
-
-void ParenCell::SetNextToDraw(Cell *next) const {
-  if (IsBrokenIntoLines())
-    m_close->SetNextToDraw(next);
-  else
-    m_nextToDraw = next;
 }